+++ src/components/Community.js
@@ -6,39 +6,9 @@ import thumbsUp from "../img/like_heart.png";
 import comment from '../img/comment.png'
 import close from '../img/close.png'
 import './Modal.css'
+import Modal from "./Modal";
 
-const Modal = ({show, onClose, comments, submitComment, newComment, setNewComment}) => {
-    return (
-        <div className={show ? "modal display-block" : "modal display-none"}>
-            <section className="modal-main">
-                <div className="modal-close-button">
-                    <img src={close} onClick={onClose} alt="Close"/>
-                </div>
 
-                <div className="comments-container">
-                    {comments.map(comment => (
-                        <div key={comment.id} className="comment">
-                            <img src={comment.profileImageUrl} alt="Profile" className="comment-profile-img"/>
-                            <div className="comment-content">
-                                <p className="comment-author">{comment.nickname}</p>
-                                <p>{comment.content}</p>
-                            </div>
-                        </div>
-                    ))}
-                </div>
-                <div className="comment-form">
-                    <input
-                        type="text"
-                        value={newComment}
-                        onChange={(e) => setNewComment(e.target.value)}
-                        placeholder="댓글을 입력해보세요."
-                    />
-                    <button onClick={submitComment}>댓글 달기</button>
-                </div>
-            </section>
-        </div>
-    );
-};
 
 const Community = () => {
     const [reviews, setReviews] = useState([]);
@@ -57,6 +27,7 @@ const Community = () => {
         setSelectedReviewId(reviewId);
         const res = await axios.get(`http://localhost:8080/api/v1/reviews/${reviewId}/comments`);
         setSelectedComments(res.data.reviewComments);
+
         setShowModal(true);
     };
 
@@ -137,21 +108,24 @@ const Community = () => {
     // 댓글 입력 처리
     const handleCommentChange = (reviewId, value) => {
         setCommentInputs(prev => ({...prev, [reviewId]: value}));
-    };
+
+    }
+
     const submitComment = async (reviewId) => {
         if (!newComment.trim()) {
             console.log("Comment is empty");
             return; // 빈 댓글은 제출하지 않음
         }
-        console.log("Submitting comment:", newComment);
+
         try {
             const memberId = 1;  // 예시 memberId, 실제로는 로그인한 사용자의 ID를 사용
             const res = await
                 axios.post(`http://localhost:8080/api/v1/reviews/${selectedReviewId}/members/${memberId}/comments`,
                     newComment
-                ,{headers:
+                    , {
+                        headers:
                             {"Content-Type": "application/json;charset=utf-8"}
-                });
+                    });
             if (res.status === 200) {
                 console.log("Comment posted successfully");
                 setNewComment("");
@@ -169,6 +143,7 @@ const Community = () => {
             const res = await axios.get(`http://localhost:8080/api/v1/reviews/${reviewId}/comments`);
             console.log(res.data)
             if (res.status === 200) {
+
                 setComments(prev => ({...prev, [reviewId]: res.data.reviewComments}));
             }
         } catch (error) {
@@ -190,7 +165,7 @@ const Community = () => {
                 <div className="reviews">
                     {reviews && reviews.length > 0 ? (
                         reviews.map((review, index) => (
-                            <div key={review.reviewId + review.memberId} className="review">
+                            <div key={index} className="review">
                                 <div className="review-header">
                                     <div key={review.member.memberId}>
                                         <div>
@@ -211,7 +186,7 @@ const Community = () => {
                                                     {review.imageUrls.map((item, idx) => (
                                                         <img
                                                             style={{display: idx < 4 ? 'block' : 'none'}}
-                                                            key={review.reviewId + " " + idx}
+                                                            key={idx}
                                                             src={item}
                                                             alt={`Review Image ${index}`}
                                                             className="review-image"/>
