+++ src/components/Modal.css
@@ -13,7 +13,7 @@
 
 .modal-main {
     background-color: white;
-    padding: 20px;
+    padding: 1%;
     width: 60%;
     height: 60%;
     border-radius: 10px;
